Fix "KeyError" exception on "System Album" and "Page" node types
Prior to this change, the Node class would throw a KeyError exception when
instantiating a "System Album" or "Page" node.  This change adds support for
the "System Album" node type (handling it the same as a regular album), and
checks the node type prior to extracting type-specific values so that unknown
types are handled gracefully.

Note there is no explicit support for the "Page" node type; this is reasonable
as the v2.0 API does not expose any actionable values for Page nodes.