Rename "album" member variable in class Node to prevent collisions
THIS IS AN INCOMPATIBLE INTERFACE CHANGE

The original code overloads the variable name "album" in three ways:
- An instantiated object (class Album);
- A string (URI), as a member variable in class Node;
- A dictionary (decoded JSON response), in the class initializer.

For example, from test.py:

	album = Album.get_album(connection, node.album)

Here, the first use of album is an object, but the last is a string.  While
the possibility of confusion (and coding errors) is real, this also will make
it harder in the future if we want to add methods that might return either an
Album object or an album URI.

This change renames the Node member variable to "album_uri", consistent with
variable naming elsewhere in the library.

This change does not change the use of the "album" name for a dictionary. It
might also be wise in the future to implement accessor methods.